Final commit

+++ Controllers/contactController.js
@@ -3,15 +3,15 @@ const Contact = require("../models/contactModel")
 
 //@desc Get all contacts
 //@route GET /api/contacts
-//access public 
+//access private 
 const getContacts = asyncHandler(async (req, res) => {
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({ user_id: req.user.id });
     res.status(200).json(contacts);
 });
 
 //@desc Create new contacts
 //@route POST /api/contacts
-//access public 
+//access private 
 const createContact = asyncHandler(async (req, res) => {
     console.log("The Reqest body is :", req.body);
     const { name, email, phone } = req.body;
@@ -23,12 +23,13 @@ const createContact = asyncHandler(async (req, res) => {
         name,
         email,
         phone,
+        user_id: req.user.id
     })
     res.status(200).json(contact);
 });
 //@desc Get single contacts
 //@route GET /api/contacts/:id
-//access public 
+//access private 
 const getContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
@@ -40,12 +41,17 @@ const getContact = asyncHandler(async (req, res) => {
 
 //@desc update single contacts
 //@route PUT /api/contacts/:id
-//access public 
+//access private 
 const updateContact = asyncHandler(async (req, res) => {
     const contact = await Contact.findById(req.params.id);
     if (!contact) {
         res.status(404);
         throw new Error("Contact not Found");
+    }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("user don't have the permission to update other user contact");
     }
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
@@ -63,6 +69,11 @@ const deleteContact = asyncHandler(async (req, res) => {
     if (!contact) {
         res.status(404);
         throw new Error("Contact not found");
+    }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("user don't have the permission to delete other user contact");
     }
     await Contact.deleteOne();
     //await Contact.remove() in  which doesn't work since this method is used to removes everything instead of the document you found.
